+++ container/cypress/e2e/wc-container.cy.js
@@ -21,6 +21,45 @@ describe('Web Container Test', () => {
         });
     });
 
+    it('getCoreSearchParams', () => {
+      const stub = cy.stub();
+
+      cy.on('window:alert', stub);
+
+      cy.get('[data-test-id="luigi-client-api-test-01"]')
+        .shadow()
+        .contains('getCoreSearchParams')
+        .click()
+        .then(() => {
+          expect(stub.getCall(0)).to.be.calledWith('{"test":"searchParam1"}');
+        });
+    });
+
+    it('getPathParams', () => {
+      const stub = cy.stub();
+      cy.on('window:alert', stub);
+
+      cy.get('[data-test-id="luigi-client-api-test-01"]')
+        .shadow()
+        .contains('getPathParams')
+        .click()
+        .then(() => {
+          expect(stub.getCall(0)).to.be.calledWith('{"path":"param"}');
+        });
+    });
+
+    it('getClientPermissions', () => {
+      const stub = cy.stub();
+      cy.on('window:alert', stub);
+      cy.get('[data-test-id="luigi-client-api-test-01"]')
+        .shadow()
+        .contains('getClientPermissions')
+        .click()
+        .then(() => {
+          expect(stub.getCall(0)).to.be.calledWith('{"permission":"testPermission"}');
+        });
+    });
+
     it('LuigiClient API getUserSettings for LuigiContainer', () => {
       const stub = cy.stub();
       cy.on('window:alert', stub);
@@ -33,9 +72,22 @@ describe('Web Container Test', () => {
           expect(stub.getCall(0)).to.be.calledWith('LuigiClient.getUserSettings()={"language":"de","date":""}');
         });
     });
+
+    it('LuigiClient API getAnchor for LuigiContainer', () => {
+      const stub = cy.stub();
+      cy.on('window:alert', stub);
+
+      cy.get('[data-test-id="luigi-client-api-test-01"]')
+        .shadow()
+        .contains('getAnchor')
+        .click()
+        .then(() => {
+          expect(stub.getCall(0)).to.be.calledWith('LuigiClient.getAnchor()="testanchor"');
+        });
+    });
   });
 
-  describe('LuigiClient API LuigiContainer', () => {
+  describe('LuigiClient API LuigiCompoundContainer', () => {
     let stub;
     beforeEach(() => {
       cy.visit('http://localhost:8080/#dashboard');
@@ -51,6 +103,17 @@ describe('Web Container Test', () => {
         .click()
         .then(() => {
           expect(stub.getCall(0)).to.be.calledWith('LuigiClient.getUserSettings()={"language":"it","date":""}');
+        });
+    });
+    it('LuigiClient API getAnchor for LuigiCompoundContainer', () => {
+      const stub = cy.stub();
+      cy.on('window:alert', stub);
+      cy.get('[data-test-id="luigi-client-api-test-compound-01"]')
+        .shadow()
+        .contains('getAnchor')
+        .click()
+        .then(() => {
+          expect(stub.getCall(0)).to.be.calledWith('LuigiClient.getAnchor()="testAnchorCompound"');
         });
     });
   });