+++ container/src/services/webcomponents.service.ts
@@ -13,23 +13,23 @@ export class WebComponentService {
   containerService: ContainerService;
   thisComponent: any;
 
-  constructor () {
+  constructor() {
     this.containerService = new ContainerService();
   }
 
-  dynamicImport (viewUrl: string) {
+  dynamicImport(viewUrl: string) {
     // Object.freeze() used as potential marker for bundlers other than webpack
     return Object.freeze(import(/* webpackIgnore: true */ viewUrl));
   }
 
-  processViewUrl (viewUrl: string, data?: any): string {
+  processViewUrl(viewUrl: string, data?: any): string {
     return viewUrl;
   }
 
   /** Creates a web component with tagname wc_id and adds it to wcItemContainer,
    * if attached to wc_container
    */
-  attachWC (
+  attachWC(
     wc_id: string,
     wcItemPlaceholder: HTMLDivElement,
     wc_container,
@@ -60,7 +60,7 @@ export class WebComponentService {
    * @param data the data to be sent
    * @param callback the callback function to be called
    */
-  dispatchLuigiEvent (msg: string, data: any, callback?: Function) {
+  dispatchLuigiEvent(msg: string, data: any, callback?: Function) {
     this.containerService.dispatch(msg, this.thisComponent, data, callback);
   }
 
@@ -72,7 +72,7 @@ export class WebComponentService {
    * @param wc_id a tagname that is used when creating the web component element
    * @returns an object with the Luigi Client API
    */
-  createClientAPI (eventBusElement, nodeId: string, wc_id: string, component: HTMLElement, isSpecialMf?: boolean) {
+  createClientAPI(eventBusElement, nodeId: string, wc_id: string, component: HTMLElement, isSpecialMf?: boolean) {
     return {
       linkManager: () => {
         return {
@@ -149,13 +149,31 @@ export class WebComponentService {
         }
         this.dispatchLuigiEvent(Events.SET_ANCHOR_LINK_REQUEST, anchor);
       },
+      getAnchor: () => {
+        return this.thisComponent.getAttribute('anchor') || '';
+      },
+      getCoreSearchParams: () => {
+        let result = this.thisComponent.getAttribute('search-params') || {};
+        result = JSON.parse(result);
+        return result;
+      },
+      getPathParams: () => {
+        let result = this.thisComponent.getAttribute('path-params') || {};
+        result = JSON.parse(result);
+        return result;
+      },
+      getClientPermissions: () => {
+        let result = this.thisComponent.getAttribute('client-permissions') || {};
+        result = JSON.parse(result);
+        return result;
+      },
       getUserSettings: () => {
         return JSON.parse(this.thisComponent.getAttribute('user-settings')) || {};
       }
     };
   }
 
-  initWC (wc: HTMLElement | any, wc_id, eventBusElement, viewUrl: string, ctx, nodeId: string, isSpecialMf?: boolean) {
+  initWC(wc: HTMLElement | any, wc_id, eventBusElement, viewUrl: string, ctx, nodeId: string, isSpecialMf?: boolean) {
     const clientAPI = this.createClientAPI(eventBusElement, nodeId, wc_id, wc, isSpecialMf);
 
     if (wc.__postProcess) {
@@ -174,9 +192,9 @@ export class WebComponentService {
    * returns a string that can be used as part of a tagname, only alphanumeric
    * characters and no whitespaces.
    */
-  generateWCId (viewUrl: string) {
+  generateWCId(viewUrl: string) {
     let charRep = '';
-    const normalizedViewUrl = new URL(viewUrl, location.href).href;
+    const normalizedViewUrl = new URL(viewUrl, encodeURI(location.href)).href;
     for (let i = 0; i < normalizedViewUrl.length; i++) {
       charRep += normalizedViewUrl.charCodeAt(i).toString(16);
     }
@@ -187,7 +205,7 @@ export class WebComponentService {
    * with the default export of the module or the first export extending HTMLElement if no default is
    * specified.
    * @returns a promise that gets resolved after successfull import */
-  registerWCFromUrl (viewUrl: string, wc_id: string) {
+  registerWCFromUrl(viewUrl: string, wc_id: string) {
     const i18nViewUrl = this.processViewUrl(viewUrl);
     return new Promise((resolve, reject) => {
       if (this.checkWCUrl(i18nViewUrl)) {
@@ -230,7 +248,7 @@ export class WebComponentService {
    * @param {*} viewUrl the source of the wc bundle
    * @param {*} onload callback function executed after script attached and loaded
    */
-  includeSelfRegisteredWCFromUrl (node, viewUrl, onload) {
+  includeSelfRegisteredWCFromUrl(node, viewUrl, onload) {
     if (this.checkWCUrl(viewUrl)) {
       /** Append reg function to luigi object if not present */
       if (!this.containerService.getContainerManager()._registerWebcomponent) {
@@ -238,7 +256,18 @@ export class WebComponentService {
           window.customElements.define(this.generateWCId(srcString), el);
         };
       }
-
+      // @ts-ignore
+      if (!window.Luigi) {
+        // @ts-ignore
+        window.Luigi = {};
+        // @ts-ignore
+        if (!window.Luigi._registerWebcomponent) {
+          // @ts-ignore
+          window.Luigi._registerWebcomponent = (src, element) => {
+            this.containerService.getContainerManager()._registerWebcomponent(src, element);
+          }
+        }
+      }
       const scriptTag = document.createElement('script');
       scriptTag.setAttribute('src', viewUrl);
       if (node.webcomponent.type === 'module') {
@@ -260,7 +289,7 @@ export class WebComponentService {
    *
    * @param {*} url the url string to check
    */
-  checkWCUrl (url: string) {
+  checkWCUrl(url: string) {
     // if (url.indexOf('://') > 0 || url.trim().indexOf('//') === 0) {
     //   const ur = new URL(url);
     //   if (ur.host === window.location.host) {
@@ -291,7 +320,7 @@ export class WebComponentService {
   /** Adds a web component defined by viewUrl to the wc_container and sets the node context.
    * If the web component is not defined yet, it gets imported.
    */
-  renderWebComponent (
+  renderWebComponent(
     viewUrl: string,
     wc_container: HTMLElement | any,
     context: any,
@@ -300,8 +329,7 @@ export class WebComponentService {
     isSpecialMf?: boolean
   ) {
     const i18nViewUrl = this.processViewUrl(viewUrl, { context });
-    const wc_id =
-      node.webcomponent && node.webcomponent.tagName ? node.webcomponent.tagName : this.generateWCId(i18nViewUrl);
+    const wc_id = node?.webcomponent?.tagName || this.generateWCId(i18nViewUrl);
     const wcItemPlaceholder = document.createElement('div');
     wc_container.appendChild(wcItemPlaceholder);
     wc_container._luigi_node = node;
@@ -338,7 +366,7 @@ export class WebComponentService {
    *
    * @param {DefaultCompoundRenderer} renderer
    */
-  createCompoundContainerAsync (renderer: any, ctx: any): Promise<HTMLElement> {
+  createCompoundContainerAsync(renderer: any, ctx: any, navNode: any): Promise<HTMLElement> {
     return new Promise((resolve, reject) => {
       // remove after review
       // if (1) {
@@ -346,18 +374,25 @@ export class WebComponentService {
       // }
       if (renderer.viewUrl) {
         try {
-          const wc_id = this.generateWCId(renderer.viewUrl);
-          this.registerWCFromUrl(renderer.viewUrl, wc_id)
-            .then(() => {
-              const wc = document.createElement(wc_id);
+          const wc_id = navNode?.webcomponent?.tagName || this.generateWCId(renderer.viewUrl);
+          const wc = document.createElement(wc_id);
+          if (navNode.webcomponent && navNode.webcomponent.selfRegistered) {
+            this.includeSelfRegisteredWCFromUrl(navNode, renderer.viewUrl, () => {
               this.initWC(wc, wc_id, wc, renderer.viewUrl, ctx, '_root');
               resolve(wc);
-            })
-            .catch(error => {
-              console.warn('Error: ', error);
-              // dispatch an error event to be handled core side
-              this.containerService.dispatch(Events.RUNTIME_ERROR_HANDLING_REQUEST, this.thisComponent, error);
             });
+          } else {
+            this.registerWCFromUrl(renderer.viewUrl, wc_id)
+              .then(() => {
+                this.initWC(wc, wc_id, wc, renderer.viewUrl, ctx, '_root');
+                resolve(wc);
+              })
+              .catch(error => {
+                console.warn('Error: ', error);
+                // dispatch an error event to be handled core side
+                this.containerService.dispatch(Events.RUNTIME_ERROR_HANDLING_REQUEST, this.thisComponent, error);
+              });
+          }
         } catch (error) {
           reject(error);
         }
@@ -375,7 +410,7 @@ export class WebComponentService {
    * @param {HTMLElement} wc_container the web component container dom element
    * @param {*} context the luigi node context
    */
-  renderWebComponentCompound (navNode, wc_container: HTMLElement, context) {
+  renderWebComponentCompound(navNode, wc_container: HTMLElement, context) {
     let renderer;
     if (navNode.webcomponent && navNode.viewUrl) {
       renderer = new DefaultCompoundRenderer();
@@ -393,8 +428,10 @@ export class WebComponentService {
 
     renderer = renderer || new DefaultCompoundRenderer();
     return new Promise(resolve => {
-      this.createCompoundContainerAsync(renderer, context)
+      this.createCompoundContainerAsync(renderer, context, navNode)
         .then((compoundCnt: HTMLElement) => {
+          (wc_container as any)._luigi_mfe_webcomponent = compoundCnt;
+          (wc_container as any)._luigi_node = navNode;
           const ebListeners = {};
           (compoundCnt as any).eventBus = {
             listeners: ebListeners,