+++ test/e2e-test-application/src/app/app.module.ts
@@ -1,7 +1,9 @@
-import { LuigiAutoNavigationService } from './services/luigi-auto-navigation.service';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
+import { LuigiAngularSupportModule } from '@luigi-project/client-support-angular';
+
+import { LuigiAutoNavigationService } from './services/luigi-auto-navigation.service';
 
 import { AppComponent } from './app.component';
 import { PreloadComponent } from './preload/preload.component';
@@ -23,7 +25,6 @@ import { ChildNode1Component } from './project/default-child/dps1/child-node-1.c
 import { ChildNode2Component } from './project/default-child/dps2/child-node-2.component';
 import { OverviewComponent } from './overview/overview.component';
 import { RestrictedComponent } from './restricted/restricted.component';
-import { LuigiContextService } from './services/luigi-context.service';
 import { GroupSettingsComponent } from './project/users/groups/group-details/group-settings/group-settings.component';
 import { DynamicComponent } from './project/dynamic/dynamic.component';
 import { HideSideNavComponent } from './project/hide-side-nav/hide-side-nav.component';
@@ -61,8 +62,8 @@ import { ViewGroupComponent } from './project/view-group/view-group.component';
     NavSyncComponent,
     ViewGroupComponent
   ],
-  imports: [BrowserModule, FormsModule, AppRoutingModule],
-  providers: [LuigiContextService, LuigiAutoNavigationService],
+  imports: [BrowserModule, FormsModule, AppRoutingModule, LuigiAngularSupportModule],
+  providers: [LuigiAutoNavigationService],
   bootstrap: [AppComponent]
 })
-export class AppModule {}
+export class AppModule { }