+++ test/e2e-test-application/src/luigi-config/extended/projectDetailNav.js
@@ -168,6 +168,7 @@ export const projectDetailNavStructure = projectId => [
       count: () => Math.floor(Math.random() * 100)
     },
     context: {
+      test: 'test',
       label: 'VirtualTree - add segments to the url',
       links: false
     },
@@ -332,6 +333,9 @@ export const projectDetailNavStructure = projectId => [
     statusBadge: {
       label: '',
       type: 'critical'
+    },
+    clientPermissions: {
+      changeCurrentLocale: false
     }
   },
   {
@@ -345,7 +349,15 @@ export const projectDetailNavStructure = projectId => [
       title: 'Hello WebComponent!'
     },
     viewUrl: '/assets/helloWorldWC.js?{i18n.currentLocale}',
-    webcomponent: true
+    webcomponent: true,
+    clientPermissions: {
+      changeCurrentLocale: true,
+      urlParameters: {
+        testParam: {
+          read: true
+        }
+      }
+    }
   },
   {
     pathSegment: 'misc2-isolated',